#Jira BRS-20 - Add pseudo folders to Xena - refactor - namespace Xena.Contracts.Domain.DocumentRelated and Xena.Web.Api.Controllers.DocumentRelated + 2nd version of migration script
