+++ src/Xena.Contracts/ApiRoutes/DocumentRoutes.cs
@@ -67,5 +67,8 @@
 
         /// <summary>"FiscalSetup/{id}/Shared"</summary>
         public const string GetSharedFromFiscal = "FiscalSetup/{id}/Shared";
+
+        /// <summary>"DocumentFolderConfiguration"</summary>
+        public const string GetDocumentFolderConfiguration = "DocumentFolderConfiguration";
     }
 }
