+++ src/Xena.Contracts/Domain/DocumentRelated/DocumentRelationDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel;
 
-namespace Xena.Contracts.Domain
+namespace Xena.Contracts.Domain.DocumentRelated
 {
     public class DocumentRelationDto : FiscalDto, IHasIdDto
     {
